CC2420Node: parameterize on flash type
