+++ java/se/sics/mspsim/platform/sentillausb/SentillaUSBNode.java
@@ -51,7 +51,7 @@ import se.sics.mspsim.platform.sky.CC2420Node;
 /**
  * Emulation of Sentilla Gateway USB Mote
  */
-public class SentillaUSBNode extends CC2420Node {
+public class SentillaUSBNode extends CC2420Node<M25P80> {
 
     public static final int MODE_LEDS_OFF = 0;
     public static final int MODE_LEDS_1 = 1;
@@ -62,7 +62,6 @@ public class SentillaUSBNode extends CC2420Node {
     public static final int GREEN_LED = 0x20;
     public static final int RED_LED = 0x10;
 
-    private final M25P80 flash;
     private SentillaUSBGui gui;
 
     private final Leds leds;
@@ -70,19 +69,13 @@ public class SentillaUSBNode extends CC2420Node {
     boolean greenLed;
 
     public SentillaUSBNode(MSP430 cpu, M25P80 flash) {
-        super("Sentilla USB", cpu);
-        this.flash = flash;
-        registry.registerComponent("xmem", flash);
+        super("Sentilla USB", cpu, flash);
         setMode(MODE_LEDS_OFF);
         leds = new Leds(cpu, LEDS);
     }
 
     public Leds getLeds() {
         return leds;
-    }
-
-    public M25P80 getFlash() {
-        return flash;
     }
 
     // USART Listener
