+++ java/se/sics/mspsim/platform/sky/MoteIVNode.java
@@ -7,7 +7,7 @@ import se.sics.mspsim.chip.SHT11;
 import se.sics.mspsim.core.IOPort;
 import se.sics.mspsim.core.MSP430;
 
-public abstract class MoteIVNode<FlashType extends ExternalFlash> extends CC2420Node {
+public abstract class MoteIVNode<FlashType extends ExternalFlash> extends CC2420Node<FlashType> {
 
   public static final int MODE_LEDS_OFF = 0;
   public static final int MODE_LEDS_1 = 1;
@@ -28,8 +28,6 @@ public abstract class MoteIVNode<FlashType extends ExternalFlash> extends CC2420
   public static final int GREEN_LED = 0x20;
   public static final int RED_LED = 0x10;
 
-  protected final FlashType flash;
-
   public boolean redLed;
   public boolean blueLed;
   public boolean greenLed;
@@ -41,18 +39,12 @@ public abstract class MoteIVNode<FlashType extends ExternalFlash> extends CC2420
   public SkyGui gui;
 
   public MoteIVNode(String id, MSP430 cpu, FlashType flash) {
-    super(id, cpu);
-    this.flash = flash;
-    registry.registerComponent("xmem", flash);
+    super(id, cpu, flash);
     setMode(MODE_LEDS_OFF);
     leds = new Leds(cpu, LEDS);
     button = new Button("Button", cpu, port2, BUTTON_PIN, true);
     sht11 = new SHT11(cpu);
     sht11.setDataPort(port1, SHT11_DATA_PIN);
-  }
-
-  public FlashType getFlash() {
-    return flash;
   }
 
   public Leds getLeds() {
